+++ pecan/tests/middleware/test_errordocument.py
@@ -1,7 +1,6 @@
 import json
 
 from webtest import TestApp
-from six import b as b_
 
 import pecan
 from pecan.middleware.errordocument import ErrorDocumentMiddleware
@@ -11,13 +10,13 @@ from pecan.tests import PecanTestCase
 
 def four_oh_four_app(environ, start_response):
     if environ['PATH_INFO'].startswith('/error'):
-        code = environ['PATH_INFO'].split('/')[2]
+        code = environ['PATH_INFO'].split('/')[2].encode('utf-8')
         start_response("200 OK", [('Content-type', 'text/plain')])
 
-        body = "Error: %s" % code
+        body = b"Error: %s" % code
         if environ['QUERY_STRING']:
-            body += "\nQS: %s" % environ['QUERY_STRING']
-        return [b_(body)]
+            body += b"\nQS: %s" % environ['QUERY_STRING'].encode('utf-8')
+        return [body]
     start_response("404 Not Found", [('Content-type', 'text/plain')])
     return []
 
@@ -33,12 +32,12 @@ class TestErrorDocumentMiddleware(PecanTestCase):
     def test_hit_error_page(self):
         r = self.app.get('/error/404')
         assert r.status_int == 200
-        assert r.body == b_('Error: 404')
+        assert r.body == b'Error: 404'
 
     def test_middleware_routes_to_404_message(self):
         r = self.app.get('/', expect_errors=True)
         assert r.status_int == 404
-        assert r.body == b_('Error: 404')
+        assert r.body == b'Error: 404'
 
     def test_error_endpoint_with_query_string(self):
         app = TestApp(RecursiveMiddleware(ErrorDocumentMiddleware(
@@ -46,7 +45,7 @@ class TestErrorDocumentMiddleware(PecanTestCase):
         )))
         r = app.get('/', expect_errors=True)
         assert r.status_int == 404
-        assert r.body == b_('Error: 404\nQS: foo=bar')
+        assert r.body == b'Error: 404\nQS: foo=bar'
 
     def test_error_with_recursion_loop(self):
         app = TestApp(RecursiveMiddleware(ErrorDocumentMiddleware(
@@ -54,8 +53,8 @@ class TestErrorDocumentMiddleware(PecanTestCase):
         )))
         r = app.get('/', expect_errors=True)
         assert r.status_int == 404
-        assert r.body == b_(
-            'Error: 404 Not Found.  (Error page could not be fetched)'
+        assert r.body == (
+            b'Error: 404 Not Found.  (Error page could not be fetched)'
         )
 
     def test_original_exception(self):