+++ pecan/tests/middleware/test_recursive.py
@@ -1,5 +1,4 @@
 from webtest import TestApp
-from six import b as b_
 
 from pecan.middleware.recursive import (RecursiveMiddleware,
                                         ForwardRequestException)
@@ -8,16 +7,16 @@ from pecan.tests import PecanTestCase
 
 def simple_app(environ, start_response):
     start_response("200 OK", [('Content-type', 'text/plain')])
-    return [b_('requested page returned')]
+    return [b'requested page returned']
 
 
 def error_docs_app(environ, start_response):
     if environ['PATH_INFO'] == '/not_found':
         start_response("404 Not found", [('Content-type', 'text/plain')])
-        return [b_('Not found')]
+        return [b'Not found']
     elif environ['PATH_INFO'] == '/error':
         start_response("200 OK", [('Content-type', 'text/plain')])
-        return [b_('Page not found')]
+        return [b'Page not found']
     elif environ['PATH_INFO'] == '/recurse':
         raise ForwardRequestException('/recurse')
     else: