+++ docs/source/changes.rst
@@ -1,3 +1,10 @@
+1.5.0
+=====
+
+* pecan no longer has a dependency on six (https://github.com/pecan/pecan/issues/144)
+* pecan now supports SQLAlchemy 2.0 (https://github.com/pecan/pecan/issues/143)
+* pecan no longer supports SQLAlchemy 1.3
+
 1.4.2
 =====